feat-10: refactor types and utils
